ml_models/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
supabase==1.0.4
python-dotenv==1.0.0
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
